+++ CHANGELOG.md
@@ -42,7 +42,10 @@
 
 - Added pie chart to sidebar view
 
-v0.1.7
+## v0.1.7
 
 - Fixed branch undefined issue
 - Colored current branch
+
+## v.1.8
+- Added autoAddToGitIgnore configuration for control add branch timer json to gitignore